Working Sample Scene
This is a working sample scene with all the setting corrected. Everything was created ground up from initial commit in master branch.

Everything that's been setup in here are:
- Player Setting
- Vuforia Component
- Sample Scene
- Camera Raycast Detection
- Item Outline
- (Unnoticeable graphical changes)

This branch has been tested with Image Target and its working "perfectly" fine, with the only known "issue" being item outline seems to be offset a bit, but doesn't affect intended goals.