Update Park Scene
Park Scene should now be complete.
I increased the range of camera raycast to 10f, just so it work nicer in reallife. Also angled the Direction Light in AR Camera just so the shadow is more visiable in the app.

The Nature Scene and Animal Scene are not functional, they only there just in case we need extra scene.